test: 增加配置项单测
