+++ packages/s2-react/src/components/switcher/constant.ts
@@ -1,6 +1,3 @@
-import { i18n } from '@antv/s2';
-import { ColIcon, RowIcon, ValueIcon } from '../icons';
-
 export const SWITCHER_PREFIX_CLS = 'switcher';
 
 export enum FieldType {
@@ -21,26 +18,3 @@ export const SWITCHER_FIELDS = [
   FieldType.Cols,
   FieldType.Values,
 ];
-
-// 是否开启行列维度相互切换
-export const getSwitcherConfig = (allowSwitchBetweenRowsAndCols = true) => ({
-  [FieldType.Rows]: {
-    text: i18n('行头'),
-    icon: RowIcon,
-    droppableType: allowSwitchBetweenRowsAndCols
-      ? DroppableType.Dimensions
-      : DroppableType.Rows,
-  },
-  [FieldType.Cols]: {
-    text: i18n('列头'),
-    icon: ColIcon,
-    droppableType: allowSwitchBetweenRowsAndCols
-      ? DroppableType.Dimensions
-      : DroppableType.Cols,
-  },
-  [FieldType.Values]: {
-    text: i18n('值'),
-    icon: ValueIcon,
-    droppableType: DroppableType.Measures,
-  },
-});
